+++ apps/dashboard/src/app/(main)/(protected)/projects/[projectId]/payments/items/page.tsx
@@ -1,0 +1,17 @@
+import PageClient from "./page-client";
+
+export const metadata = {
+  title: "Items",
+};
+
+type Params = {
+  projectId: string,
+};
+
+export default async function Page({ params }: { params: Promise<Params> }) {
+  return (
+    <PageClient />
+  );
+}
+
+