+++ packages/stack-shared/src/config/schema.ts
@@ -5,10 +5,10 @@ import { DEFAULT_EMAIL_TEMPLATES, DEFAULT_EMAIL_THEMES, DEFAULT_EMAIL_THEME_ID }
 import * as schemaFields from "../schema-fields";
 import { offerSchema, userSpecifiedIdSchema, yupBoolean, yupDate, yupMixed, yupNever, yupNumber, yupObject, yupRecord, yupString, yupTuple, yupUnion } from "../schema-fields";
 import { isShallowEqual } from "../utils/arrays";
-import { SUPPORTED_CURRENCIES } from "../utils/currencies";
+import { SUPPORTED_CURRENCIES } from "../utils/currency-constants";
 import { StackAssertionError } from "../utils/errors";
 import { allProviders } from "../utils/oauth";
-import { DeepFilterUndefined, DeepMerge, DeepRequiredOrUndefined, deleteKey, filterUndefined, get, has, isObjectLike, mapValues, set, typedAssign, typedFromEntries } from "../utils/objects";
+import { DeepFilterUndefined, DeepMerge, DeepRequiredOrUndefined, deleteKey, filterUndefined, get, has, isObjectLike, mapValues, set, typedAssign, typedEntries, typedFromEntries } from "../utils/objects";
 import { Result } from "../utils/results";
 import { CollapseObjectUnion, Expand, IntersectAll, IsUnion, typeAssert, typeAssertExtends, typeAssertIs } from "../utils/types";
 import { Config, NormalizationError, NormalizesTo, assertNormalized, getInvalidConfigReason, normalize } from "./format";
@@ -114,18 +114,15 @@ const branchAuthSchema = yupObject({
 });
 
 export const branchPaymentsSchema = yupObject({
-  stripeAccountId: yupString().optional(),
-  stripeAccountSetupComplete: yupBoolean().optional(),
   autoPay: yupObject({
     interval: schemaFields.dayIntervalSchema,
   }).optional(),
-  exclusivityGroups: yupRecord(
-    userSpecifiedIdSchema("exclusivityGroupId"),
-    yupRecord(
-      userSpecifiedIdSchema("offerId"),
-      yupBoolean().isTrue(),
-    ),
-  ),
+  groups: yupRecord(
+    userSpecifiedIdSchema("groupId"),
+    yupObject({
+      displayName: yupString().optional(),
+    }),
+  ).meta({ openapiField: { description: 'The groups that offers can be in. All offers in a group (besides add-ons) are mutually exclusive.', exampleValue: { "group-id": { displayName: "My Group" } } } }),
   offers: yupRecord(
     userSpecifiedIdSchema("offerId"),
     offerSchema,
@@ -135,13 +132,6 @@ export const branchPaymentsSchema = yupObject({
     yupObject({
       displayName: yupString().optional(),
       customerType: schemaFields.customerTypeSchema,
-      default: yupObject({
-        quantity: yupNumber(),
-        repeat: schemaFields.dayIntervalOrNeverSchema.optional(),
-        expires: yupString().oneOf(['never', 'when-repeated']).optional(),
-      }).default({
-        quantity: 0,
-      }),
     }),
   ),
 });
@@ -213,8 +203,8 @@ export const environmentConfigSchema = branchConfigSchema.concat(yupObject({
     trustedDomains: yupRecord(
       userSpecifiedIdSchema("trustedDomainId"),
       yupObject({
-        baseUrl: schemaFields.wildcardUrlSchema,
-        handlerPath: schemaFields.handlerPathSchema,
+        baseUrl: schemaFields.wildcardUrlSchema.max(300),
+        handlerPath: schemaFields.handlerPathSchema.max(300),
       }),
     ),
   })),
@@ -248,7 +238,7 @@ export function migrateConfigOverride(type: "project" | "branch" | "environment"
 
   // BEGIN 2025-07-28: domains.trustedDomains can no longer be an array
   if (isEnvironmentOrHigher) {
-    res = mapProperty(res, "domains.trustedDomains", (value) => {
+    res = mapProperty(res, p => p.join(".") === "domains.trustedDomains", (value) => {
       if (Array.isArray(value)) {
         return typedFromEntries(value.map((v, i) => [`${i}`, v]));
       }
@@ -259,14 +249,27 @@ export function migrateConfigOverride(type: "project" | "branch" | "environment"
 
   // BEGIN 2025-07-28: themeList and templateList have been renamed (this was before the email release, so they're safe to remove)
   if (isBranchOrHigher) {
-    res = removeProperty(res, "emails.themeList");
-    res = removeProperty(res, "emails.templateList");
+    res = removeProperty(res, p => p.join(".") === "emails.themeList");
+    res = removeProperty(res, p => p.join(".") === "emails.templateList");
   }
   // END
 
   // BEGIN 2025-07-28: sourceOfTruth was mistakenly written to the environment config in some cases, so let's remove it
   if (type === "environment") {
-    res = removeProperty(res, "sourceOfTruth");
+    res = removeProperty(res, p => p.join(".") === "sourceOfTruth");
+  }
+  // END
+
+  // BEGIN 2025-08-25: stripeAccountId and stripeAccountSetupComplete are unused, so let's remove them
+  if (type === "environment") {
+    res = removeProperty(res, p => p.join(".") === "payments.stripeAccountId");
+    res = removeProperty(res, p => p.join(".") === "payments.stripeAccountSetupComplete");
+  }
+  // END
+
+  // BEGIN 2025-08-25: payments.items.default is no longer used, so let's remove it
+  if (isBranchOrHigher) {
+    res = removeProperty(res, p => p.length === 4 && p[0] === "payments" && p[1] === "items" && p[3] === "default");
   }
   // END
 
@@ -274,41 +277,41 @@ export function migrateConfigOverride(type: "project" | "branch" | "environment"
   return res;
 };
 
-function removeProperty(obj: any, path: string): any {
-  return mapProperty(obj, path, () => undefined);
+function removeProperty(obj: Record<string, any>, pathCond: (path: (string | symbol)[]) => boolean): any {
+  return mapProperty(obj, pathCond, () => undefined);
 }
 
-function mapProperty(obj: any, path: string, mapper: (value: any) => any): any {
-  const keyParts = path.split(".");
-
-  for (let i = 0; i < keyParts.length; i++) {
-    const pathPrefix = keyParts.slice(0, i).join(".");
-    const pathSuffix = keyParts.slice(i).join(".");
-    if (has(obj, pathPrefix) && isObjectLike(get(obj, pathPrefix))) {
-      const newValue = mapProperty(get(obj, pathPrefix), pathSuffix, mapper);
-      set(obj, pathPrefix, newValue);
-    }
-  }
-  if (has(obj, path)) {
-    const newValue = mapper(get(obj, path));
-    if (newValue !== undefined) {
-      set(obj, path, newValue);
+function mapProperty(obj: Record<string, any>, pathCond: (path: string[]) => boolean, mapper: (value: any) => any): any {
+  const res: Record<string, any> = Array.isArray(obj) ? [] : {};
+  for (const [key, value] of typedEntries(obj)) {
+    const path = key.split(".");
+    if (pathCond(path)) {
+      const newValue = mapper(value);
+      if (newValue !== undefined) {
+        set(res, key, newValue);
+      } else {
+        // do nothing
+      }
+    } else if (isObjectLike(value)) {
+      set(res, key, mapProperty(value, p => pathCond([...path, ...p]), mapper));
     } else {
-      deleteKey(obj, path);
+      set(res, key, value);
     }
   }
-
-  return obj;
+  return res;
 }
 import.meta.vitest?.test("mapProperty - basic property mapping", ({ expect }) => {
-  expect(mapProperty({ a: { b: { c: 1 } } }, "a.b.c", (value) => value + 1)).toEqual({ a: { b: { c: 2 } } });
-  expect(mapProperty({ a: { b: { c: 1 } } }, "a.b.d", (value) => value + 1)).toEqual({ a: { b: { c: 1 } } });
-  expect(mapProperty({ x: 5 }, "x", (value) => value * 2)).toEqual({ x: 10 });
-  expect(mapProperty({ a: { b: { c: 1 } } }, "b.c", (value) => value * 10)).toEqual({ a: { b: { c: 1 } } });
-  expect(mapProperty({ a: 1 }, "b.c", (value) => value)).toEqual({ a: 1 });
+  expect(mapProperty({ a: { b: { c: 1 } } }, p => p.join(".") === "a.b.c", (value) => value + 1)).toEqual({ a: { b: { c: 2 } } });
+  expect(mapProperty({ a: { b: { c: 1 } } }, p => p.join(".") === "a.b.d", (value) => value + 1)).toEqual({ a: { b: { c: 1 } } });
+  expect(mapProperty({ x: 5 }, p => p.join(".") === "x", (value) => value * 2)).toEqual({ x: 10 });
+  expect(mapProperty({ a: { b: { c: 1 } } }, p => p.join(".") === "b.c", (value) => value * 10)).toEqual({ a: { b: { c: 1 } } });
+  expect(mapProperty({ a: 1 }, p => p.join(".") === "b.c", (value) => value)).toEqual({ a: 1 });
+  expect(mapProperty({ "a.b": { c: 1 } }, p => p.join(".") === "a.b.c", (value) => value + 1)).toEqual({ "a.b": { c: 2 } });
+
+  expect(mapProperty({ a: { b: { c: 1 } } }, p => p.length === 3 && p[0] === "a" && p[1] === "b", (value) => value + 1)).toEqual({ a: { b: { c: 2 } } });
 });
 
-function renameProperty(obj: any, oldPath: string, newPath: string): any {
+function renameProperty(obj: Record<string, any>, oldPath: string, newPath: string): any {
   const oldKeyParts = oldPath.split(".");
   const newKeyParts = newPath.split(".");
   if (!isShallowEqual(oldKeyParts.slice(0, -1), newKeyParts.slice(0, -1))) throw new StackAssertionError(`oldPath and newPath must have the same prefix. Provided: ${oldPath} and ${newPath}`);
@@ -444,16 +447,18 @@ const organizationConfigDefaults = {
   },
 
   payments: {
-    stripeAccountId: undefined,
-    stripeAccountSetupComplete: false,
     autoPay: undefined,
-    exclusivityGroups: (key: string) => (key: string) => undefined,
+    groups: (key: string) => ({
+      displayName: undefined,
+    }),
     offers: (key: string) => ({
       displayName: key,
-      customerType: undefined,
+      groupId: undefined,
+      customerType: "user",
       freeTrial: undefined,
       serverOnly: false,
       stackable: undefined,
+      isAddOnTo: false,
       prices: (key: string) => ({
         ...typedFromEntries(SUPPORTED_CURRENCIES.map(currency => [currency.code, undefined])),
         interval: undefined,
@@ -461,18 +466,14 @@ const organizationConfigDefaults = {
         freeTrial: undefined,
       }),
       includedItems: (key: string) => ({
-        quantity: undefined,
+        quantity: 0,
         repeat: "never",
         expires: "when-repeated",
       }),
     } as const),
     items: (key: string) => ({
       displayName: key,
-      default: {
-        quantity: 0,
-        expires: "when-repeated",
-        repeat: "never",
-      },
+      customerType: "user",
     } as const),
   },
 } as const satisfies DefaultsType<OrganizationRenderedConfigBeforeDefaults, [typeof environmentConfigDefaults, typeof branchConfigDefaults, typeof projectConfigDefaults]>;
@@ -652,6 +653,22 @@ export async function sanitizeOrganizationConfig(config: OrganizationRenderedCon
     ...DEFAULT_EMAIL_TEMPLATES,
     ...prepared.emails.templates,
   };
+  const offers = typedFromEntries(typedEntries(prepared.payments.offers).map(([key, offer]) => {
+    const isAddOnTo = offer.isAddOnTo === false ?
+      false as const :
+      typedFromEntries(Object.keys(offer.isAddOnTo).map((key) => [key, true as const]));
+    const prices = offer.prices === "include-by-default" ?
+      "include-by-default" as const :
+      typedFromEntries(typedEntries(offer.prices).map(([key, value]) => {
+        const data = { serverOnly: false, ...(value ?? {}) };
+        return [key, data];
+      }));
+    return [key, {
+      ...offer,
+      isAddOnTo,
+      prices,
+    }];
+  }));
   return {
     ...prepared,
     emails: {
@@ -660,8 +677,13 @@ export async function sanitizeOrganizationConfig(config: OrganizationRenderedCon
       themes,
       templates,
     },
+    payments: {
+      ...prepared.payments,
+      offers
+    }
   };
 }
+
 
 /**
  * Does not require a base config, and hence solely relies on the override itself to validate the config. If it returns
@@ -877,6 +899,7 @@ type BranchConfigNormalizedOverride = Expand<ValidatedToHaveNoConfigOverrideErro
 type EnvironmentConfigNormalizedOverride = Expand<ValidatedToHaveNoConfigOverrideErrors<typeof environmentConfigSchema>>;
 type OrganizationConfigNormalizedOverride = Expand<ValidatedToHaveNoConfigOverrideErrors<typeof organizationConfigSchema>>;
 
+
 // Overrides
 // ex.: { a?: null | { b?: null | number, c: string }, d?: null | number, "a.b"?: number, "a.c"?: string }
 export type ProjectConfigOverride = NormalizesTo<ProjectConfigNormalizedOverride>;
@@ -899,6 +922,7 @@ export type BranchIncompleteConfig = Expand<ProjectIncompleteConfig & BranchConf
 export type EnvironmentIncompleteConfig = Expand<BranchIncompleteConfig & EnvironmentConfigNormalizedOverride>;
 export type OrganizationIncompleteConfig = Expand<EnvironmentIncompleteConfig & OrganizationConfigNormalizedOverride>;
 
+
 // Rendered configs before defaults, normalization, and sanitization
 type ProjectRenderedConfigBeforeDefaults = Omit<ProjectIncompleteConfig,
   | keyof BranchConfigNormalizedOverride
@@ -913,6 +937,7 @@ type EnvironmentRenderedConfigBeforeDefaults = Omit<EnvironmentIncompleteConfig,
   | keyof OrganizationConfigNormalizedOverride
 >;
 type OrganizationRenderedConfigBeforeDefaults = OrganizationIncompleteConfig;
+
 
 // Rendered configs before sanitization
 type ProjectRenderedConfigBeforeSanitization = Expand<Awaited<ReturnType<typeof applyProjectDefaults<ProjectRenderedConfigBeforeDefaults>>>>;