+++ packages/template/src/lib/stack-app/apps/interfaces/admin-app.ts
@@ -34,6 +34,7 @@ export type StackAdminApp<HasTokenStore extends boolean = boolean, ProjectId ext
   & AsyncStoreProperty<"emailThemes", [], { id: string, displayName: string }[], true>
   & AsyncStoreProperty<"emailPreview", [{ themeId?: string | null | false, themeTsxSource?: string, templateId?: string, templateTsxSource?: string }], string, false>
   & AsyncStoreProperty<"emailTemplates", [], { id: string, displayName: string, themeId?: string, tsxSource: string }[], true>
+  & AsyncStoreProperty<"stripeAccountInfo", [], { account_id: string, charges_enabled: boolean, details_submitted: boolean, payouts_enabled: boolean } | null, false>
   & {
     useEmailTemplates(): { id: string, displayName: string, tsxSource: string }[], // THIS_LINE_PLATFORM react-like
     listEmailTemplates(): Promise<{ id: string, displayName: string, tsxSource: string }[]>,
@@ -76,8 +77,12 @@ export type StackAdminApp<HasTokenStore extends boolean = boolean, ProjectId ext
 
     setupPayments(): Promise<{ url: string }>,
     createStripeWidgetAccountSession(): Promise<{ client_secret: string }>,
-    createPurchaseUrl(options: { customerId: string, offerId: string }): Promise<string>,
-    createItemQuantityChange(options: { customerId: string, itemId: string, quantity: number, expiresAt?: string, description?: string }): Promise<void>,
+    createItemQuantityChange(options: (
+      { userId: string, itemId: string, quantity: number, expiresAt?: string, description?: string } |
+      { teamId: string, itemId: string, quantity: number, expiresAt?: string, description?: string } |
+      { customCustomerId: string, itemId: string, quantity: number, expiresAt?: string, description?: string }
+    )): Promise<void>,
+    testModePurchase(options: { priceId: string, fullCode: string, quantity?: number }): Promise<void>,
   }
   & StackServerApp<HasTokenStore, ProjectId>
 );