+++ apps/e2e/CHANGELOG.md
@@ -1,5 +1,23 @@
 # @stackframe/e2e-tests
 
+## 2.8.35
+
+### Patch Changes
+
+- Various changes
+- Updated dependencies
+  - @stackframe/stack-shared@2.8.35
+  - @stackframe/js@2.8.35
+
+## 2.8.34
+
+### Patch Changes
+
+- Various changes
+- Updated dependencies
+  - @stackframe/stack-shared@2.8.34
+  - @stackframe/js@2.8.34
+
 ## 2.8.33
 
 ### Patch Changes