+++ examples/demo/CHANGELOG.md
@@ -1,5 +1,24 @@
 # @stackframe/example-demo-app
 
+## 2.8.35
+
+### Patch Changes
+
+- Updated dependencies
+  - @stackframe/stack-shared@2.8.35
+  - @stackframe/stack@2.8.35
+  - @stackframe/stack-ui@2.8.35
+
+## 2.8.34
+
+### Patch Changes
+
+- Various changes
+- Updated dependencies
+  - @stackframe/stack-shared@2.8.34
+  - @stackframe/stack@2.8.34
+  - @stackframe/stack-ui@2.8.34
+
 ## 2.8.33
 
 ### Patch Changes