+++ examples/docs-examples/CHANGELOG.md
@@ -1,5 +1,23 @@
 # @stackframe/docs-examples
 
+## 2.8.35
+
+### Patch Changes
+
+- Updated dependencies
+  - @stackframe/stack-shared@2.8.35
+  - @stackframe/stack@2.8.35
+  - @stackframe/stack-ui@2.8.35
+
+## 2.8.34
+
+### Patch Changes
+
+- Updated dependencies
+  - @stackframe/stack-shared@2.8.34
+  - @stackframe/stack@2.8.34
+  - @stackframe/stack-ui@2.8.34
+
 ## 2.8.33
 
 ### Patch Changes