+++ examples/react-example/CHANGELOG.md
@@ -1,5 +1,17 @@
 # react-example
 
+## 2.8.35
+
+### Patch Changes
+
+- @stackframe/react@2.8.35
+
+## 2.8.34
+
+### Patch Changes
+
+- @stackframe/react@2.8.34
+
 ## 2.8.33
 
 ### Patch Changes